refactor: share OS-thread pool for contract compile
