+++ chain/chain/Cargo.toml
@@ -19,6 +19,7 @@ enum-map.workspace = true
 itertools.workspace = true
 lru.workspace = true
 num-rational.workspace = true
+num_cpus.workspace = true
 parking_lot.workspace = true
 rand_chacha.workspace = true
 rand.workspace = true
@@ -27,7 +28,6 @@ serde.workspace = true
 strum.workspace = true
 tempfile.workspace = true
 thiserror.workspace = true
-thread-priority.workspace = true
 time.workspace = true
 tokio.workspace = true
 tracing.workspace = true
