+++ core/async/Cargo.toml
@@ -18,6 +18,7 @@ parking_lot.workspace = true
 rand = { workspace = true, optional = true }
 serde.workspace = true
 serde_json.workspace = true
+thread-priority.workspace = true
 time.workspace = true
 tokio = { workspace = true, features = ["rt", "rt-multi-thread", "macros"] }
 tokio-util.workspace = true
